Write decent route test
