+++ requirements.txt
@@ -14,8 +14,7 @@ Flask-User==1.0.1.5
 pytest==3.0.5
 pytest-cov==2.4.0
 
-
-# Google API 
+# Google API
 google-api-python-client==1.6.2
 google-auth==0.8.0
 google-auth-httplib2==0.0.2
